:recycle: added AvroSerializer
