+++ src/main/java/org/kafkahq/repositories/SchemaRegistryRepository.java
@@ -3,8 +3,8 @@ package org.kafkahq.repositories;
 import io.confluent.kafka.schemaregistry.client.rest.entities.requests.ConfigUpdateRequest;
 import io.confluent.kafka.schemaregistry.client.rest.exceptions.RestClientException;
 import io.confluent.kafka.serializers.KafkaAvroDeserializer;
-import io.confluent.kafka.serializers.KafkaAvroSerializer;
 import org.kafkahq.models.Schema;
+import org.kafkahq.modules.AvroSerializer;
 import org.kafkahq.modules.KafkaModule;
 import org.kafkahq.utils.PagedList;
 import org.kafkahq.utils.Pagination;
@@ -23,7 +23,7 @@ public class SchemaRegistryRepository extends AbstractRepository {
     @Inject
     private KafkaModule kafkaModule;
     private Map<String, KafkaAvroDeserializer> kafkaAvroDeserializers = new HashMap<>();
-    private Map<String, KafkaAvroSerializer> kafkaAvroSerializers = new HashMap<>();
+    private AvroSerializer avroSerializer;
 
     public PagedList<Schema> list(String clusterId, Pagination pagination, Optional<String> search) throws IOException, RestClientException, ExecutionException, InterruptedException {
         return PagedList.of(all(clusterId, search), pagination, list -> list
@@ -192,5 +192,12 @@ public class SchemaRegistryRepository extends AbstractRepository {
         }
 
         return this.kafkaAvroDeserializers.get(clusterId);
+    }
+
+    public AvroSerializer getAvroSerializer(String clusterId) {
+        if(this.avroSerializer == null){
+            this.avroSerializer = new AvroSerializer(this.kafkaModule.getRegistryClient(clusterId));
+        }
+        return this.avroSerializer;
     }
 }
